+++ toys-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+### 0.3.3 / TBD
+
+* CHANGED: Renamed file_utils helper to fileutils.
+* ADDED: Highline helper
+* ADDED: Spinner helper
+
 ### 0.3.2 / 2018-05-07
 
 * CHANGED: Split core engine out into "toys-core" from the "toys" gem.